requirements.txt
streamlit>=1.28.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy
arxiv>=2.0.0
wikipedia-api
PyPDF2>=3.0.0
python-dotenv>=1.0.0
transformers>=4.36.0
requests>=2.28.0
torch
PyMuPDF
wikipedia>=1.4.0